gradle 2.1 binaries are applied
Gradle 2.1 enhanced api in a way to allow task cancellation.
We need that for the ticket below:
https://code.google.com/p/android/issues/detail?id=59464

Gradle v.2.1 has already been applied to the IJ 'master' branch
(before branching for 14.x). JetBrains guys say that they will not
backport it to 13.0/1.x.

Current CL is a combination of the following jetbrains commits:
https://github.com/JetBrains/intellij-community/commit/9843325487ab334d1b9e824d35b70284f949a800
https://github.com/JetBrains/intellij-community/commit/1fd58e13c9968b79ad4c91a3c6ed0197efc346c8
https://github.com/JetBrains/intellij-community/commit/0b21a62636323525e668f0a95214ed1b2de68a6b

Plus it adds a couple of custom changes.

Change-Id: Ib8cd253982c3f6b57cfaa938631ecf4bf68ded66