Fixed logic that updates Gradle version in wrapper before syncing.
The old logic was checking that the URL started with 'http' while
in reality it starts with 'https'. The fix involved existing method
in GradleUtil to get the current Gradle version in wrapper file.

This CL also sets the right distribution type (DEFAULT_WRAPPED)
before checking (minor fix.)

This CL may fix
https://code.google.com/p/android/issues/detail?id=75075
but since I haven't been able to reproduce the issue, I'm not closing
the bug until the filer confirms that the issue is no longer present.

Change-Id: I86dc11faac5b8f8e9426038791fc0ed2267aa66c