Fixed error handling of missing methods in settings.gradle.
When there is a missing method (e.g. a typo like 'incude' instead
of 'include') the IDE shows the error but the quick-fixes are not
applicable to the error:
- Open Gradle Settings
- Apply Gradle plug-in
In addition, the location of the error is missing, clicking on the
error does not take the user to the error location.

This CL fixes this by not showing irrelevant quick-fixes and
including the error location.

Fixes
https://code.google.com/p/android/issues/detail?id=76732

Change-Id: I83be2281eb88c36d770a4d469bf33e2a0bef4327