Fix for: Nav editor: getting some weird rendering of menu
https://code.google.com/p/android/issues/detail?id=76927

I couldn't reliably reproduce this but have seen the same problem 
come up intermittently in local projects. It could be timing related.

There's certainly a bug in the caching optimization in the the 
component that renders Activities and Menus in the Navigation Editor.

This CL fixes that bug, by having the rendering operation reset the cache
(as it should).

Change-Id: I20798b80e0befc8ae41c229c73bed6f5e26da821