Avoid needlessly rendering layout editor in XML mode
I just noticed that the layout editor is performing rendering
operations when you are in the XML editor editing layouts. This is
completely unnecessary work which can slow down the IDE.

This seems to be the result of the new support for preferring the XML
editor for layouts; the changed editor initialization order means that
the layout editor is created and then hidden.

However, during creation the layout editor would unconditionally
register a PSI listener and treat it as activated. This is redundant
(and incorrect) since the PSI listener will be activated when the
editor itself is activated.

Change-Id: I1e6997b7cf16500a358dc23d84f7c91452cecafa
(cherry picked from commit 55b5c7256fdd07457677c41c257fb85844fc3637)