Now we catch NoSuchMethodError in addition to UnsupportedMethodException.
In class IdeaJavaProject we invoke GradleProject#getBuildDirectory,
a method that was introduced in Gradle 2.0. We invoke this method in
a try-catch block, in case a project is using Gradle 1.x. We were
catching UnsupportedMethodException, but a user has reported that
we need to catch NoSuchMethodError too. This CL fixes that method
invocation, and other places where we are catching only
UnsupportedMethodException.

Fixes
https://code.google.com/p/android/issues/detail?id=77060

Change-Id: I0378eacd441f99846df8ce9d721696fba0f200de