Added hint to error message when sync fails due to class loading.
User reported that this known issue can happen when using an old
version of JDK 1.7 (user has using 1.7.0_10). He also reported
that using the latest one (1.7.0_67) solves the problem.

This CL adds a hint, and a hyperlink to open the "SDKs" panel in
the "Project Structure" dialog if the project is using JDK 1.7.

Fixes:
https://code.google.com/p/android/issues/detail?id=76360

Change-Id: I10d10129b88ee483fafd3192d90db190ac617a9e