Now we offer to automatically install platform if missing android.jar.
Before this CL, if syncing failed because the platform being used in
the project is missing android.jar (or just failed loading) we just
showed a generic message and a hyperlink to open the SDK manager.

Now we provide the cause of the error (e.g. android.jar missing) and
a quick-fix to automatically install the platform and sync the
project.

This is related to this issue:
https://code.google.com/p/android/issues/detail?id=75934

I haven't been able to reproduce exactly what Tor experienced, but
now, with this CL, we provide a better user experience if android.jar
is missing.

Change-Id: Id84fcd49a38f8110c59393e2d3af5b5cba1578b7