Support visible paths with non-visible steps.
This change removes an exception when opening a wizard with first
path not having any visible steps. This use case is important when
we need to run path wizard completion actions without showing any
steps associated with the path (e.g. if we got the information
elsewhere, as in the case of installer data handoff)

Following changes were made:
1. DynamicWizard#init will no longer attempt to show step from the
   first path, it will now share implementation with the "Next"
   button.
2. Wizard state will now be stabilized before init happens and UI
   is shown. This allows for accurate computation of the visibility
   and validation state.
3. (Implementation detail) Dynamic wizard and its paths will now share
   update queue. This reduces the latency between update request
   submission and execution (and allows easier implementation of item #2)

Change-Id: I77d23a840230c0aa82c84f05eadc0ee6101114f4