Now provide better hints if sync fails when problems with JDKs are detected.
Gradle is trying to reuse a daemon that is using a different SDK than the one
specified in Android Studio. This CL provides a better message by telling the
user the path of the expected JDK and the path being used, and offers a
hyperlink to open the Project structure dialog.

Fixes
https://code.google.com/p/android/issues/detail?id=76984

Change-Id: If48110a2e8d23dd15fd6f204146dafbd50543473