Log output from emulator to console
This CL wires up the emulator output to a console in the run tool
window. This allows users to see any errors reported by the emulator.

This is achieved by introducing a new class ExternalToolRunner. Given
a command line, creates a process by executing that command line, and
attaches that process to a console. The console includes a single
stop application action in its toolbar that can be used to kill the
running process.

Fixes https://code.google.com/p/android/issues/detail?id=75905

Change-Id: I428e3670b0f6ecdb9836e1c13f871ac6c4dd50ae