Fix scaling of device diagram.
Previously each device diagram was reduced in size by 1/2 until
the size was below some threshold. In a single category (e.g.
"Phones") it was possible to have a larger device needing to be
scaled more times than a smaller one, resulting in the diagram for
the larger device showing as smaller than the smaller device.
Now all devices are scaled based on what is necessary for the
largest device.
https://code.google.com/p/android/issues/detail?id=77635

Change-Id: I7e784d8dc173a6cd3706873aa1a9a2969a6e603b