navigator: Show modules with no source roots
The Android Project View defaulted to not showing modules that
don't have any source roots. This was done partially to not show the
empty root module that is common in projects created by the new project
wizard.

This CL changes the implementation so that only the root level module
is not shown if it doesn't have any source roots. Modules at all other
levels are always shown whether they have any source roots or not..

Fix for https://code.google.com/p/android/issues/detail?id=77595

Change-Id: Iab1081cf5316bd5fa9160cc11eb3e10886780b8f