75050: Improved support for RTL attributes.
align_startOf and similar attributes weren't being taken into account
when calculating the dependecy graph for the components in the relative
layout. That was probably causing the problem in 75050.

Improved support for the designer editor when using an RTL language.
Currently, when editing and using an RTL language the components will
snap to the opposite side of where they are dropped (if dropped on the
right side of the layout, it will snap to the left side because of the
wrong attributes being set).
Also added support for margin and layout_parent RTL attributes.

Change-Id: I69a39dd0f977239421091dafa848723ba3a8916b