Improved algorithm for automatic layout of states in Navigation View
This CL simply sorts the states according to the number of children they have,
minus the number of parents they have. This tends to put the base activities on
the left side of the graph and leaves on the right. This produces a good layout
for ioshed and the other apps in each of their different form factors.

Also bug fixes for problems with uniqueness of states and misplaced test
for whether a node had already been visited.

Change-Id: Idaf2769af06d78af0923569710c07d724f710590