Keep hidden version of the Android SDK repo
We do not maintain Android SDK version so there is no good way for the Android Studio
first run wizard to know if it should update the SDK. This change makes the wizard
store the "version" of the Android SDK downloaded by the wizard the last time.

This version number will be tied to Android Studio release.

Change-Id: I9a4701d1ffc9ecf1a56c8f0fdcfd1623c1da20f0