Making the avd name consistent throughout the ui
When creating a new AVD the user can specify a display name to be
used in the AVD manager. Another name is also generated based on
the display name that is used in command-line tools, and as the
name for generated files. Previously these were inconsistently used
in the UI. This change adds:
- Optionally showing the generated name in the AVD configuration
  screen.
- Updating the generated name when the display name is changed.
- Showing the display name rather than the generated name in the
  run/debug dialog, in the combo box and the list of running
  emulators.
https://code.google.com/p/android/issues/detail?id=78040
https://code.google.com/p/android/issues/detail?id=78033

Change-Id: Id0fcbf0ca2e111d2cd0c1891e471472f30febc1e