Simplify event propagation in Navigation Editor
Belt and braces for (fixed) issue:

https://code.google.com/p/android/issues/detail?id=78042

This bug was caused an event storm which, although fixed by

https://android-review.googlesource.com/#/c/111772/

was indicative of an architectural flaw which this CL addresses more directly.

Don't trigger code generation by listening to changes in the NavigationModel.
Instead, have the NavigationView call the code generator directly and let
the NavigationModel refresh as a result of those changes.

Change-Id: I79daf6b110eba9c22226545699286e46aad6c2a2