Gradle task can't be stopped
https://code.google.com/p/android/issues/detail?id=59464

This ticket targets rather complex problem. The first step was
to apply gradle 2.1 binaries as they provide an api for gradle
tasks cancellation.

After that we found out that there is a problem with gradle
tasks execution when android plugin is enabled -
AndroidGradleTaskManager intercepts task execution request
and performs the job in a separate thread, thus letting IDE
execution UI look in a way like the task has already finished

This CL fixes that problem - it enhances GradleTasksExecutor
API in a way to allow to wait for the actual task execution and
forwards task output to the IDE run/debug tool window.

Corresponding test has been added as well (IDE fixtures
framework is expanded)

The next step will be to support actual task cancellation.

Change-Id: Ie9d9eaafac158d495232f5f12df2e277a1c0dc2e