Setting sd card property when one is specified.
Right now whether a device has an sd card (hw.sdCard=yes in 
config.ini) is determined by whether there is any removable
storage specified in the device config (regardless of what's 
in the avd config). This change causes the specification of
a sd card size or file to imply that the hardware supports 
sd cards.
https://code.google.com/p/android/issues/detail?id=78338

Change-Id: Ic1502d7c80da0583a544a9455a7a92a00b580204