Removed unnecessary children in 'JDK' node in Project View.
By default all jars in the JDK node in the Project View are displayed,
which adds noise and can be confusing to Android developers. For example,
they may see the JavaFX jar and think that it is available in Android.

This CL cleans that up. It removes all the children in the JDK node
except for rt.jar. In rt.jar all children are removed, except for the
packages 'java' and 'javax', like this:

- External Libraries
  - JDK
    - rt.jar
      - java
      - javax

This change is purely cosmetic. The real bug is that we should not use
JDK at all for Android modules, and resolve 'java' and 'javax' classes
against the Android SDK. Fixing this bug requires deeper changes in
IDEA itself (there are few other complications, e.g. JDK needed by
Java library projects.) We'll tackle this bug after 1.0. The bug is
being tracked here:
https://code.google.com/p/android/issues/detail?id=77028

Change-Id: Ie88f97bfd2da42b27f0726995a34f1537aa29d42