Fixed usability issues with "License Agreement" wizard step.
This is the wizard where users accept/decline license before
installing a SDK component. The problems were:

The dialog has 2 panels, 1 for the components to install, and
the other for the license text. The dialog did not have a splitter
and the panel that contained the license text was too narrow to
be read comfortably. In addition the Decline/Accept buttons were
set in a way that the the 2 panels did not look properly aligned.

This CL adds a splitter and aligns the panels. It uses IntelliJ's
Splitter, which is the component IDEA uses everywhere. The
downside is that it is not possible to use that component with
the UI designer. I removed the UI designer form and built the
panel programmatically (the panel is really simple).

Change-Id: I8d996fb3bc31500e4f6686ee5a417b228f35dc69