78826: Templates use wrong parent class
Adds a new "hasDependency" Freemarker method which can be used by the
templates to check whether a given dependency is used by the project
the template is being instantiated into.

This lets the appcompat-dependent activity templates correctly pick
the right parent activity to use.

This also switches the "should we use appcompat?" logic for new
projects over from "is minSdkVersion < 14?" to "is minSdkVersion < 21
and compileSdkVersion >= 21" since this is now the recommended way to
handle Material Design for apps.

The hasDependency code uses the new getResolvedCoordinate method in
the Gradle model, which can be used to accurately check the set of
dependencies. Lint's code predated that API and used various hacky
heuristics (and the support-v4 check probably doesn't work anymore now
that support-v4 is shipped as an .aar instead of a .jar), so I
replaced it with calls to the same utility method now used by the
template dependency lookup.

Change-Id: I83711b372cd69866384accf7775a580dae6e020b