78072: Fix invalid SDK detection
We have some code to detect if the Android SDK is valid, but it had
a couple of problems:
(1) it was only calld on load synced model from disk, and
(2) it only checked whether all SDK entries exist, not that android.jar
    was actually among them, and
(3) the solution was to reinstall the whole platform, which isn't always
    necessary (e.g. when the SDK configuration is missing classpath entries
    but they're actually there on disk)

We now look to see if the SDK looks incorrect, and if so, we recreate
all the SDK roots. Only if that does not help do we continue with the
SDK reinstallation.

Change-Id: I0e14e3a688c5be2451f91faac3982d315056ed5f