79243: Emulator skin doesn't work for round device
First, round clipping does not work with "Use Host GPU" enabled, so we
turn that flag off by default specifically for round devices. If a
developer is deliberately trying to emulate a round device, that's
probably more important than acceleration. (And with the small screen
resolution of the wear devices, lack of Host GPU is less severe than
say for a tablet AVD.)

Second, the emulator skin didn't work right for round clipping. Rather
than trying to figure it out, I just copied in the round skin from the
system image:
  system-images/android-20/android-wear/x86/skins/AndroidWearRound
along with the config files - since the round skin used in the layout
renderer isn't for a specific device anyway.

Change-Id: If88ebf2747bd0e94b269553656a63512efe922c0