studio.exe error message on 64-bit Windows
Users on 64-bit Windows machines should not be using studio.exe;
instead, they should be using studio64.exe. However, if they do
that now, they currently see a misleading JVM error message
which is only a side-effect of the real problem.

This CL fixes that case, now showing the user an error message
that tells them to run the correct program to proceed.

Change-Id: I533e6a7f9dd27402bcd7a0e39db880c734c11a3d