Resolve class names before matching them.
[This CL is intended to be cherry picked into 1.1]

Rework the unification (matching) code the Navigation Editor 
to use fully qualified class names and resolve simple(r) names 
according to import statements. Also rework all the templates 
to use fully qualified names so that they can be resolved without 
a context (reference to import statements). 

This is to fix the problem below. 

The fragment:

    x.setOnClickListener(new View.OnClickListener() {...})

doesn't match in the LoginActivity example because LoginActivity uses:

    import android.view.View.OnClickListener;

    x.setOnClickListener(new OnClickListener() {...})

instead.

Fully resolve class names before comparing them so that we aren't dependent 
on import style.

Change-Id: Ic350b3ed8deeafce3a3116422e89bc0a88e97c38