80099: Add a missing check for START segment type.
The assertion is caused because a segment type START is not detected in
the ConstraintPainter (the END case was fixed in
https://android-review.googlesource.com/#/c/111706/).

This also addresses the assertion error in 80098. isLeftSegment was
@NotNull but there are cases where null can be passed so simply changed
Nullable (since it's safe to do so).

Change-Id: I5f1063eb53cb03b51766d87b8343f1a2a9854b93
(cherry picked from commit b8d8dea982af7df92593a93e8adbcd15722332fd)