Fixed bug where it was impossible to change the path of the JDK in the IDE.
The bug was introduced by this CL
https://android-review.googlesource.com/113941 . After this CL was
submitted, the IDE will delete and recreate a JDK when the path is
changed. It turns out that one step was missing: set the new JDK in
the project. Because this was not being done, after project sync
the IDE was trying to guess the best IDE for the project's language
level, and that guess was not always reliable. This CL also removes
that guess.

Fixes:
https://code.google.com/p/android/issues/detail?id=80698

Change-Id: I2d7fbdac2cdcf53e133ea2fe0b169d7e572ab0a7
(cherry picked from commit 6459cb0c9aabb36859d3ca7f42d266c26590dbe9)