80100: Reduce String creation on AndroidValueResourcesIndex
AndroidValueResourcesIndex#normalizeDelimiters was a source of a huge
number of String objects in the heap. Most of the time we don't need to
create a new String, the original String is already in the "normalized"
format, so only create new Strings when we really need to.

Change-Id: I9aea708e0c7bacdcdbaa28f76d0fb829442e86c4
(cherry picked from commit 24761f775aa19cf8e8a1773e7bf6e08ce748bf63)