Welcome Wizard now reads Unicode handoff file directly
Java property files come with a bunch of special
rules that we have to be careful to follow. Since the
installer currently doesn't do any of that, for now,
we just revert to loading the file manually and
parsing it ourselves.

Change-Id: I1655957a6f88029abfe33af505e6505ff500f412
(cherry picked from commit 40a3ecb9df49bb64b6be27a4f3bbfdae74023073)