device chooser: special case Google APIs target
Projects could be compiled against addon platforms that provide
optional libraries at run time. e.g. if you set:
   compileSdkVersion "Google Inc.:Google APIs:21"
then you can assume that 3 additional libraries provided by the
Google API add-on will be available at runtime.

The device chooser used to conservatively display that it is
unsure if such a project can be run on a device since it doesn't
know whether the device actually has that library.

This CL special cases that check just for the Google APIs addon,
assuming that Google APIs are available on all devices (this is
not true for devices w/o access to Play?), and doesn't say that
it is unsure. Instead, it'll report the device as compatible.

Note that this is purely informational: the device chooser will
allow you to run the app on that device regardless of this CL.

Fix for https://code.google.com/p/android/issues/detail?id=81499

Change-Id: I004c959c087eb1c2405749dd108abf366d830338