apk installation: do not throw error for unexpected dumpsys output
After installing an apk on a device, we perform a "dumpsys package"
on the installed package to obtain the installation timestamp. This
was used to identify whether an apk needs to be installed again in
the future.

Unexpected output from the dumpsys command would result in an error
being logged. This CL changes the severity of that log to a warning.

Fix for https://code.google.com/p/android/issues/detail?id=79778

Change-Id: Ia4575a9872ce0b2ca4126df5e8b0004ef2604c29
(cherry picked from commit 2a201dc63a3e88e8257c4d1815fff50e5d45a602)