Now we restart the IDE when stopping Gradle daemons.
Context:
When we get one of those weird ClassNotFoundExceptions, we provide
a "quick fix" hyperlink to stop all Gradle daemons. The problem is
that the API that Gradle provides to stop daemons does much more
than just stopping daemons: it stops the whole Tooling API, making
it impossible for the IDE to use Gradle.

I talked to the Gradle folks in person, and they confirmed this is
the intended behavior. For the IDE to become usable again, a restart
is required.

Now, what this CL does is the following:
- If the IDE supports restart, we stop the daemons and restart the IDE (we warn users
  about this behavior)
- If the IDE does not support restart, we offer a
  hyperlink to the docs that describe how to stop Gradle daemons)

I thought about just removing the hyperlink, but stopping the daemons
can be handy when getting this type of errors.

Fixes:
https://code.google.com/p/android/issues/detail?id=80591

Change-Id: If2d7ae251ed0065a49c6763fad871cdd64187a41
(cherry picked from commit 3eaed7142ece2239cba162e6582d184560d34c86)