device chooser: Prefer to use a device rather than the emulator
When the device chooser is displayed, and no devices are connected,
it selects the option to launch a new emulator. Often, users connect
a device when they see this but don't remember to select the pick
device option..So this CL changes its behavior to always select
the pick device option whenever it detects that a device is available.

Fix for https://code.google.com/p/android/issues/detail?id=81166

Change-Id: Iade56a3a9f401e3450a5321069f0c27d2dc4b1b4