Force stop package before running a previously installed version.
When launching an application, we don't re-install the application
if we detect that there have been no changes made locally. As a result,
running an application will simply bring it to the foreground if the
latest version was already running on the device.

Some users have come to rely on running an application having the
effect of restarting the application rather than just bringing it to
the foreground. In order to achieve that behavior, this CL force
stops the application in such a case before starting it, thereby
allowing it to go through the fully activity startup life cycle.

Fixes https://code.google.com/p/android/issues/detail?id=81163

Change-Id: I9f7c69fd22ca0c41f090b77908b088812101704e