Adding defaults for new avd creation wizard
Devices are hardcoded, since it's not clear exactly when we will
want to switch from one to another (e.g. right now we don't want
to use the Nexus 6 since it's more powerful than the "average"
device nowadays).
The default system image is one with the highest version. If an x86
image with the highest version exists, it is chosen. If not, an
image with the highest version is chosen arbitrarily.
Fixes https://code.google.com/p/android/issues/detail?id=82503

Change-Id: If3a5e23b94b65540d853a6aebd020492beeeebd9