Removes annotations.jar from SDKs used by project (if necessary)
When a project is opened, we perform an additional check: if the
IDEA Android SDK has annotations.jar and API level > 15, we
recreate the SDK without annotations.jar.

Context: On Windows, SDK tools cannot be updated because the
file annotations.jar is being locked by the IDE. For most projects,
annotations.jar should not have been there in the first place;
it was supposed to only be added if compileSdkVersion <= 15.

For this CL to work, it needs also
https://android-review.googlesource.com/119813

Change-Id: I1765cb87927ff68df65e94b148b136cb1b96f84c