82813: uncommitted fragment transactions not highlighted by lint
The CleanupDetector in lint has been rewritten from a bytecode
analyzer to an AST analyzer, so register it as an inspection
in the IDE. Also implement equals() on the ResolvedNode classes
which is required by the new detector to correctly check variable
identity.

Also removes some deprecated calls with the new recommended
calls.

Change-Id: I772b023fa6a5d590f3225cb65a8db2b6846ae160