Upgraded installer to NSIS3+ [DO NOT MERGE]
Upgrading was simple:
- Convert our script to UTF8
- Add a Unicode identifier which NSIS3 uses

Previously, we were using a forked Unicode version of NSIS to
support our installer. Though not developed by Nullsoft, it was
endorsed by them, and it served as a useful stopgap solution
for many projects for the past few years. NSIS only recently
released NSIS3 which provides Unicode support (and is still in
Beta), so it seemed reasonable to base our script on the
forked project rather than NSIS3.

However, since a goal of ours now is to build NSIS on Linux if
possible, it looks like the official NSIS distribution is much
more likely to provide this than the forked project ever was, as
they didn't make doing so a priority.

I also took advantage of this update to add a README as well as
package all non-standard components that we rely on, since this
upgrade forced me to start from a new NSIS installation, letting
me see exactly what was necessary to make our script compile and
run.

Change-Id: Id0236a27f9744c59d9126d3d5ee7dcf0f1d7f884
(cherry picked from commit f955b97fb6d1630b95e3f28dc280f9770ddd2eb3)