Adding checkbox to enable/disable AVD skin
There will be a new entry in the avd ini file to specify
the skin that was previously selected, if any. That way if
the user (without going into advanced mode) unchecks the box,
saves, and then comes back an checks it again we can restore
the skin that was previously selected. Otherwise the observable
configuration would be the same, but the behavior different.
Fixes https://code.google.com/p/android/issues/detail?id=82742

Change-Id: I1b693416db48e321e6057683262c64ed44bf69f3