Improved selection of JDKs in PSD on Mac.
Selecting a JDK on Mac, using the 'Project Structure'
dialog (PSD), requires the user to go to
${JDK_HOME}/Contents/Home which is not obvious, and
just selecting ${JDK_HOME} from the file chooser results
in a 'invalid JDK' error.

In this CL, when a user selects ${JDK_HOME}, the PSD
will check if the selected directory has /Contents/Home
in it and if it does, the PSD will use that full path as
the JDK location to store.

Change-Id: Ib03ebc7598b65cda86582fd60086d8540c233c2a