Device editor fixes
The clone/edit button is based on the whether the action is active,
but the condition is only actually evaulated when when action is
changed. The condition starts as false (as initially nothing is
selcted), but as long as only unmodifiable devices are clicked on
the button will never become enabled, since the action is never
changed.

Selecting the new/edited device after creating/editing.

Adding the ability to create wear/tv hardware profiles.

Adjusting appearance of device edit page to align more closely
with avd edit page, and switching to SingleStepWizard to remove
unnecessary back/next buttons and a little bit of code.
Change-Id: I96b94f993660c4d2aa9176f532a31eac16364f75