Improved "Open File" quick-fix when using wrong JDK version.
This "quick-fix" shows up when using JDK 1.6 with android-21
compileSdkVersion. Before, the quick-fix will just open the
build.gradle file, positioning the cursor at the first character
of the file, which is pretty useless.

This CL fixes the quick-fix by finding the position of the
compileSdkVersion *value*, to make it easier to user to change
such value (e.g. from 21 to 19).

Change-Id: I0884cc8b2ac5074cfc7d4c4017b0c475008722a2