We no longer manually remove modules after a project sync.
Problem: it was not possible to include modules using
conditions in settings.gradle files. Such conditionally-added
modules never showed up in the project.

Solution: it turns out that this is a bug in Android
Studio and not Gradle. Gradle correctly imports all modules,
but after project sync, the IDE removes all modules that
it thinks are not in settings.gradle file, using pretty
basic parsing, which does not understand conditions or
custom code. This CL removes that module cleanup code.

Fixes
https://code.google.com/p/android/issues/detail?id=77158

Change-Id: Id1db66726bb68fb1ea6828362bb9b35abd1b95e5