Support for construction of master-detail apps from nav/layout editors
The last pieces of infrastructure needed to support the construction of
master-detail apps from within the Navigation Editor . Previous versions
parsed the idioms of the master-detail template app (these will continue
to be supported) but could not create these sort of apps from scratch.

This CL allows the Navigation Editor to add listeners to Fragments
within activities and so allows the user to break an Activity into
named fragments, whose behavior can vary for different configurations.

The assembly of Activities from Fragments is done through the
Layout Editor in the usual way.

Change-Id: I0b90f6cc08bb3dbb184c2a838e5085b61a5224bc