Fixing resource merge to work within new project wizard
When run during the new project wizard, no project is set in the
template. A project is required for any Psi file usage, so using
the default project in that case.
Switching to using the string format of the target xml, since
if we're using the default project VirtualFile lookups don't work.
This caused the new code to no longer be auto-formatted,
so also adding some basic format-preserving functionality.
Fixes https://code.google.com/p/android/issues/detail?id=98317

Change-Id: Icf9c55b386a2800c44cf8e8f7bd5692605414f11