Fix NPE when starting the theme editor.
The theme editor was relying on the file passed to be part of the
project. Since it's a virtual file seems that getting AndroidFacet
fails.
We do not really need the file, since we can use the module obtained
from the configuration and remove one unnecessary parameter.

Also fix an out of bounds exception when switching to simple mode
while editing a cell.

Update theme editor after the libraries have finished loading. This will
solve a problem with AppCompat themes not being displayed if the theme
editor was being restored from a previous session.

Change-Id: I2ecba6dd59b3758887e6686a87c40a29cf26f336