Allow user select the modules to import.
This feature is useful for developers working on
large Android applications (e.g. 50+ modules) by
reducing the number of modules that are visible
in the IDE. By reducing the number of modules,
IDE functionality (source indexing, compiling) has
better performance.

This feature is *disabled* by default. Users can enable
it by going to the Preferences/Settings dialog and
navigate to the "Gradle > Experimental" page.

This CL:
- Presents the user a dialog with the list of modules
  of the project being imported. Users can select
  which modules to include in the project.
- The dialog offers "speed search" (search as you
  type) to make it easy to find modules in the list.
- The dialog allows users to save module selection
  to a disk and load module selection from disk.
- Subsequent project syncs respects the original module
  selection.
- Inter-module dependencies support. If a module is
  excluded, dependent modules are linked to the compiled
  version of the excluded module.
- Module selection is persisted between IDE sessions.
- Includes UI tests.

Change-Id: I38dff4465bfe478d0c7be3ee71295c935f841b0f