BCP 47 support: Use LocaleQualifiers
Also revise the device configuration panel such that you are
editing only a single locale qualifier (and make it easier to
pick related languages). Also gets rid of the old Add/Edit
Translations actions from the layout rendering menu and replaces
them with an action to open the strings editor in the module.

Change-Id: I9bd71025f1e3cfab839c25b13f3c6254ec2b77bf