Reduce class churn from layout rendering
Up until now, each new render request would create a new class loader
which would load custom views and R classes. (This was also done for
measure requests, which for example is done during interactive
dragging in a LinearLayout to figure out the intrinsic sizes of
objects for weight calculations.)

Unfortunately, on some JVMs, these classes are *not* properly garbage
collected, which means they accumulate - and after a while, the IDE
runs out of permgen space, which effectively renders the IDE unusable.

This CL attempts to greatly reduce this problem, by reusing the
rendering class loader to a much greater extent. It now keeps track of
the .class files it has loaded and the file modified timestamps at the
loading time. Then, when a new render class loader is requested, it
checks whether any of the loaded files have changed, and if not, it
simply reuses the previous class loader. (It also makes sure the
layout library, which is used as the parent class loader, has not
changed.)

This is not a perfect fix, but is deliberately small to avoid big
changes right before 1.1: it still uses a single class loader for all
loaded classes including classes from .jar files. In a followup CL I
will revise this where we use a hierarchy of loaders: a separate
loader for each module, a separate loader for each set of .jar file
libraries etc, such that we can replace only the class loaders for
local .class files and continue to reuse the previous class loader for
non-changing dependencies like guava.jar.

It also doesn't have a great place to hang the class loaders off of,
so it uses a weak hashmap in the project class loader itself for now.
In the follow-up CL, I will split the RenderService up into two parts:
a persistent per-module RenderService, and temporary RenderTasks which
is what most of the current RenderService currently is. When that's
done, the ClassLoader caching will be handled by the RenderService,
and the ProjectClassLoader will be split up into a class loader
hierarchy.

I also split up the
is-java-source-file-modified-after-last-compilation check: Rather than
having that check in the class loader, it's now in the view loader,
such that it isn't only run on first class load but on each class use,
since otherwise caching would prevent this check from being performed
on each render. That also allowed me to move the logger and credential
fields out of the class loader, which was important to allow the class
loader to be shared between multiple separate render tasks.

I left some class loading logging in place (turned off by a static
flag) for now, while we tweak this behavior.

Change-Id: I63ff678a0f536c806734519f5c57537c022f8677
(cherry picked from commit 2d18e0542d35b1592384d1b3379e32794b2934a0)