Fixes dragging tabs out of the android tool window.
The android tool window uses a "RunnerUI" which is only
meant to run when a build is being "run". Runner UI has
a hidden dependency on the content manager of the execution
manager. If it has not been created then trying to create
a container for a "dragged" tab crashes.

Change-Id: I7b2f0e06d81d86549bd00f109a741c6e266cf3cb