Removed obsolete interfaces in Project Structure
The project structure dialog had two interfaces,
ConfigurableHost and ValidationAwareConfigurable, which look like
they either were intended to be used, or were used at some point
but not any longer.

The thought for now is, let's clean up the interfaces, and if we
end up needing either of them later, we can add them again at that
time.

Change-Id: I9266afa29408db92c6d55388a5a72904874caba8