Android Tool Window: First step towards more tabs
In order to have more tabs that monitor aspects of the device
the device/app selection was removed from being a tab and moved
at the top of the tool window. This way is clear that the
selection of device/app affect all the tabs of the tool window.
This is the first step in moving the memory monitor tool window
as a tab of the android tool window.

Change-Id: I89731efb91ccd5f1124e93e0e06ba396c000dea4