strings editor: Show missing locales in sorted order
When translations are missing for a set of locales, the error message
reported the sequence of missing locales in a random order. This CL
changes that sequence to be sorted by the locale label. This fix is
primarily to make sure that unit tests don't rely on an unknown
ordering.

Change-Id: Ie91020e2b8b18e7f1d1bd9f36334d30a7365d0d4