Rename RenderService to RenderTask
This CL is just a pure rename refactoring from
RenderService to RenderTask (along with variable name
changes from renderService to renderTask, etc.)

Similarly, the ProjectClassLoader is renamed to
ModuleClassLoader, to reflect the fact that it is in
fact module specific.

There are no semantic changes.

(This is in preparation for a followup CL which will
make structural changes.)

Change-Id: I27d1830499f973565fa0ca493c47d6f428207845